+++ src/features/klant/contactverzoek/ContactverzoekenOverzicht.vue
@@ -11,31 +11,6 @@
         </contactmoment-preview>
       </template>
     </contactverzoeken-overzicht>
-
-    <!-- 
-    <template v-if="records.length">
-      <thead>
-        <tr>
-          <th>Tel. nummer</th>
-          <th>E-mailadres</th>
-          <th class="row-link-header">Details</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr v-for="(record, i) in records" :key="i" class="row-link">
-          <td>
-            <pre>{{ record }}</pre>
-          </td>
-          <td>{{ record.emailadres }}</td>
-          <td>
-            <router-link
-              :to="`/contactverzoeken/${record.uuid}`"
-              title="Details"
-            />
-          </td>
-        </tr>
-      </tbody>
-    </template> -->
   </table>
 </template>
