+++ src/views/ContactenverzoekenView.vue
@@ -6,6 +6,6 @@
 </template>
 
 <script setup lang="ts">
-import ContactverzoekenZoeker from "@/features/klant/contact/ContactverzoekenZoeker.vue";
+import ContactverzoekenZoeker from "@/features/klant/contactverzoek/ContactverzoekenZoeker.vue";
 import { Heading as UtrechtHeading } from "@utrecht/component-library-vue";
 </script>
